requirements.txt
annotated-types==0.7.0
blis==1.0.1
Bottleneck==1.4.2
Brotli==1.0.9
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.10
einops==0.8.0
en_core_web_sm==3.8.0
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.10.0
ftfy==6.3.1
hausdorff==0.2.6
huggingface_hub==0.24.6
idna==3.7
importlib-metadata==7.0.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lightning-utilities==0.11.9
llvmlite==0.43.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
monai==1.4.0
mpmath==1.3.0
murmurhash==1.0.11
networkx==3.4.2
numba==0.60.0
numexpr==2.10.1
numpy==1.24.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==11.0.0
pip==24.2
preshed==3.0.9
ptflops==0.7.4
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.0
sentence-transformers==3.3.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.3
thinc==8.3.2
threadpoolctl==3.5.0
timm==1.0.11
tokenizers==0.20.3
torch==2.3.1+cu118
torchmetrics==1.6.0
torchvision==0.18.1+cu118
tqdm==4.66.5
transformers==4.46.3
triton==2.3.1
typer==0.13.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wheel==0.44.0
wrapt==1.17.0
zipp==3.20.2
